# specs/001-azure-devops-timetracker/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SMIT TimeTrack API
  version: 0.1.0-draft
  description: >-
    Contrato inicial. O esquema de autenticação é provisório até conclusão da prova
    técnica documentada em research.md. Não implementar este bearer como token
    Azure DevOps sem validação de emissor, audiência e organização.
servers:
  - url: https://api.example.invalid/v1
security:
  - BackendSession: []
components:
  securitySchemes:
    BackendSession:
      type: http
      scheme: bearer
      description: Sessão curta emitida pelo backend após troca verificável de identidade; desenho pendente.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
  schemas:
...
    Timer:
      type: object
      description: >-
        `id` é opaco (identificador interno do backend), não um GUID do
        Azure DevOps — só `projectId`/`workItemId` vêm do Azure DevOps.
      required: [id, workItemId, startedAtUtc, status]
      properties:
        id: { type: string }
        workItemId: { type: integer, minimum: 1 }
        startedAtUtc: { type: string, format: date-time }
        status: { type: string, enum: [active, stopped] }
        activityTypeId: { type: [string, 'null'] }
    Entry:
      type: object
      required: [id, workItemId, localDate, timezone, durationSeconds, source, billable]
      properties:
        id: { type: string }
        workItemId: { type: integer, minimum: 1 }
        localDate: { type: string, format: date }
        timezone: { type: string, example: America/Sao_Paulo }
        durationSeconds: { type: integer, minimum: 1 }
        startTime: { type: [string, 'null'], example: '09:00', description: 'Hora local de início (HH:MM); nulo se o lançamento não tem horário.' }
        endTime: { type: [string, 'null'], example: '10:30', description: 'Início + duração, em hora local.' }
        source: { type: string, enum: [timer, manual] }
        billable: { type: boolean }
        activityTypeId: { type: string }
        note: { type: string, maxLength: 2000 }
        revision: { type: integer, minimum: 1 }
...
paths:
...
  /me/timer:
    get:
      operationId: getActiveTimer
      responses:
        '200':
          description: Timer atual ou null.
          content:
            application/json:
              schema:
                oneOf:
                  - { $ref: '#/components/schemas/Timer' }
                  - { type: 'null' }
        '401': { description: Não autenticado }
    post:
      operationId: startTimer
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [projectId, projectName, workItemId]
              properties:
                projectId: { type: string, format: uuid, description: GUID do projeto no Azure DevOps. }
                projectName: { type: string, description: Nome do projeto no momento da chamada (afirmação do cliente; ver WorkItemAccessService). }
                workItemId: { type: integer, minimum: 1 }
                activityTypeId: { type: string, description: ID interno de um tipo de atividade do tenant. }
                title: { type: string, description: 'Título do work item, lido do form já aberto — cosmético.' }
                workItemType: { type: string, description: 'Tipo do work item, lido do form já aberto — cosmético.' }
      responses:
        '201':
          description: Timer iniciado.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Timer' }
        '403': { description: Projeto/item inacessível }
        '409': { description: Já existe timer ativo para outro item }
  /me/timer/stop:
    post:
      operationId: stopTimer
      parameters:
        - { $ref: '#/components/parameters/IdempotencyKey' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [timerId]
              properties:
                timerId: { type: string }
                note: { type: string, maxLength: 2000 }
                billable: { type: boolean }
                overtimeReason: { type: string, maxLength: 500, description: 'Perfil restrito: motivo do trecho fora do expediente sem pedido aprovado.' }
                overtimeAcknowledged: { type: boolean, description: 'Perfil restrito: a pessoa marcou que entendeu que a hora só conta se for confirmada.' }
      responses:
        '200':
          description: >-
            Timer fechado; uma ou mais fatias diárias geradas. Perfil restrito: o
            trecho fora do expediente sem pedido aprovado não vira lançamento, vira
            hora extra a confirmar; o cabeçalho `X-Pending-Overtime` diz quantas.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Entry' }
        '409': { description: Timer já fechado por outra operação ou semana bloqueada }
        '422': { description: 'Perfil restrito sem `overtimeReason`/`overtimeAcknowledged` (o timer continua ativo)' }
...